backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
alembic==1.14.0
pydantic==2.10.3
pydantic-settings==2.7.0
playwright==1.49.1
google-genai==1.14.0
beautifulsoup4==4.12.3
httpx==0.28.1
